Refactor, test, customtemplatetag
Refactored the view to only using one service function.

While writing the test for it I encountered a strange bug and investigated. Found that the old code was overriding the value of the time_account_balance to the std min sek format instead of putting the value in.

First tried to work on the function and found that this was the wrong aproch.

Instead created a new custom template tag that does the same as the function and changes the format.

Todo next is to also do the same for the total so that the dict that contains the data for the view only has the int = sek values and the conversion is only happening in the view. That way if on a later date the format is supposed to be changed it is only at one place.