added new model parameter to sessions, added new function to sessions admin
Now when the admin view for sessions is opened all sessions are checked if they are by bots. If the duration of the session was 0 it will be declared that of a bot.

I also added the starttime to the view so it is ordered by time and it is possible to see when people visited