Added single bot file per plattform. Now instead folderstructure forms each bot.
Just add a new command to the commands.py file and it will automaticly be added to the bot.

Removed need for hardcoded bot_register_ids

This makes development easier.

Now I can have a dev version of the customer bot with one ID localy for testing and development and when the new version is deployed to the server it will automaticly use the id there. No conflict between dev and prod version.