added static ngrok domain
Ngrok added one free static domain to their fee user package.

This helps during development if a webservice needs to send messages to the local dev version.

added to settings a env in the list of allowed hosts

added to the env list NGROK_STATIC_DOMAIN for a placeholder for a domain